refactored daily into components and style doc
