+++ client/src/components/styles/Daily.styled.js
@@ -1,0 +1,36 @@
+import styled from "styled-components";
+import {
+  StyledHabit,
+  HabitLogo,
+  HabitInfo,
+  HabitName,
+  HabitStreakContainer,
+  HabitStreak,
+} from "./Habit.styled";
+
+export const StyledDaily = styled(StyledHabit)``;
+
+export const DailyLogo = styled(HabitLogo)``;
+
+export const DailyInfo = styled(HabitInfo)``;
+
+export const DailyName = styled(HabitName)``;
+
+export const DailyStreakContainer = styled(HabitStreakContainer)``;
+
+export const DailyStreak = styled(HabitStreak)``;
+
+export const Control = styled.div`
+  border-radius: 5px 0 0 5px;
+  background: #50b5e9;
+  padding: 15px 6px 30px 6px;
+  box-shadow: 0 1px 3px rgb(0 0 0 / 0.35);
+`;
+
+export const ControlButton = styled.button`
+  border-radius: 3px;
+  background: #ffffff80;
+  cursor: pointer;
+  width: 28px;
+  height: 28px;
+`;
