+++ client/src/components/styles/UserTasks.styled.js
@@ -98,136 +98,8 @@ export const QuickAdd = styled.textarea`
   border: none;
 `;
 
-export const Control = styled.div`
-  border-radius: ${(props) => (props.left ? "5px 0 0 5px" : "0 5px 5px 0")};
-  background: ${(props) => {
-    switch (props.$type) {
-      case "habit":
-        return "#24CC8F";
-      case "daily":
-        return "#50B5E9";
-      case "todo":
-        return "#FFBE5D";
-      default:
-        return "#EDECEE";
-    }
-  }};
-
-  padding: 15px 6px 30px 6px;
-  box-shadow: 0 1px 3px rgb(0 0 0 / 0.35);
-`;
-
-export const ControlButton = styled.button`
-  border-radius: ${(props) => (props.habit ? "50%" : "3px")};
-  background: ${(props) => (props.habit ? "#1A181D40" : "#FFFFFF80")};
-  cursor: pointer;
-  width: 28px;
-  height: 28px;
-`;
-
-// Habits
-export const Habit = styled.div`
-  display: flex;
-  margin: 0 0 3px 0;
-`;
-
-export const HabitLogo = styled.img`
-  width: 10px;
-  height: 10px;
-`;
-
-export const HabitInfo = styled.div`
-  background: #ffffff;
-  width: 100%;
-  padding: 5px 0 0 0;
-  display: flex;
-  flex-direction: column;
-`;
-
-export const HabitName = styled.span`
-  color: #34313a;
-  font-size: 16px;
-  font-family: "Roboto", sans-serif;
-  padding: 4px 0 10px 12px;
-`;
-
-export const HabitStreakContainer = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  margin: 4px 0 0;
-  padding: 0 8px;
-`;
-
-export const HabitStreak = styled.span`
-  color: #a5a1ac;
-  font-size: 12px;
-  padding: 0 5px;
-`;
-
-// Dailies
-export const Daily = styled(Habit)``;
-
-export const DailyLogo = styled(HabitLogo)``;
-
-export const DailyInfo = styled(HabitInfo)``;
-
-export const DailyName = styled(HabitName)``;
-
-export const DailyStreakContainer = styled(HabitStreakContainer)``;
-
-export const DailyStreak = styled(HabitStreak)``;
-
-// To Do's
-export const ToDo = styled(Habit)``;
-
-export const ToDoInfo = styled(HabitInfo)``;
-
-export const ToDoName = styled(HabitName)``;
-
-// Rewards
-export const Item = styled.div`
-  width: 94px;
-  height: 120px;
-  border-radius: 4px 4px 0 0;
-  margin: 10px 0 0 0;
-  background: #ffffff;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
 export const ItemsContainer = styled.div`
   display: flex;
   flex-direction: row;
   justify-content: space-evenly;
-`;
-
-export const ItemPriceContainer = styled.div`
-  width: 94px;
-  height: 28px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background: #ffbe5d33;
-  margin: 0 1px;
-  border-radius: 0 0 4px 4px;
-  box-shadow: 0 1px 3px rgb(0 0 0 / 0.35);
-`;
-
-export const ItemPrice = styled.span`
-  font-size: 12px;
-  color: #794b00;
-  font-weight: bold;
-`;
-
-export const ItemPriceLogo = styled(TagsLogo)`
-  margin: 0 5px 0 0;
-`;
-
-export const ItemImage = styled.div`
-  width: 68px;
-  height: 68px;
-  background: #dddddd;
-  margin: 12px 0;
 `;
